chore: Separate BuildingController and BuildingAdminController
